requirements.txt
pandas==2.2.3
plotly==5.24.1
prefect==3.3.1
python-dotenv==1.1.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10